requirements.txt
asgiref==3.8.1
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
dnspython==2.6.1
PyJWT==2.8.0
pymongo==4.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
